# swagger.yaml
openapi: 3.0.0
info:
  title: Maze Runner Books API
  version: 1.0.0
  description: API para gestión de libros de Maze Runner - James Dashner

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://p3-32057736.onrender.com
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    Product:
      type: object
      required:
        - name
        - description
        - price
        - isbn
        - publicationYear
        - publisher
        - pageCount
        - coverType
        - genre
        - ageRange
        - categoryId
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        isbn:
          type: string
        publicationYear:
          type: integer
        publisher:
          type: string
        edition:
          type: string
        language:
          type: string
        pageCount:
          type: integer
        coverType:
          type: string
          enum: [Hardcover, Paperback, E-book]
        genre:
          type: string
        ageRange:
          type: string
        categoryId:
          type: integer
        tagIds:
          type: array
          items:
            type: integer
    
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string

tags:
  - name: Public - Products
    description: Public product endpoints
  - name: Admin - Products
    description: Admin product management endpoints
  - name: Admin - Categories
    description: Category management endpoints
  - name: Admin - Tags
    description: Tag management endpoints
  - name: Authentication
    description: User authentication endpoints

paths:
  /products:
    get:
      tags: [Public - Products]
      summary: Search and filter products
      description: Public endpoint for searching and filtering Maze Runner books with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search in name and description
        - in: query
          name: category
          schema:
            type: string
          description: Category ID or name
        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated tag IDs
        - in: query
          name: price_min
          schema:
            type: number
          description: Minimum price
        - in: query
          name: price_max
          schema:
            type: number
          description: Maximum price
        - in: query
          name: publisher
          schema:
            type: string
          description: Filter by publisher
        - in: query
          name: coverType
          schema:
            type: string
            enum: [Hardcover, Paperback, E-book]
          description: Filter by cover type
        - in: query
          name: language
          schema:
            type: string
          description: Filter by language
        - in: query
          name: genre
          schema:
            type: string
          description: Filter by genre
        - in: query
          name: ageRange
          schema:
            type: string
          description: Filter by age range
        - in: query
          name: publicationYear
          schema:
            type: integer
          description: Filter by publication year
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer

    post:
      tags: [Admin - Products]
      summary: Create a new product
      description: Create a new Maze Runner book (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: Product created successfully
        401:
          description: Unauthorized - Invalid or missing token

  /products/{id}:
    get:
      tags: [Admin - Products]
      summary: Get product by ID
      description: Get a specific product by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Product ID
      responses:
        200:
          description: Product found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Product'
        404:
          description: Product not found
        401:
          description: Unauthorized

    put:
      tags: [Admin - Products]
      summary: Update a product
      description: Update an existing product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        200:
          description: Product updated successfully
        404:
          description: Product not found
        401:
          description: Unauthorized

    delete:
      tags: [Admin - Products]
      summary: Delete a product
      description: Delete a product (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Product ID
      responses:
        200:
          description: Product deleted successfully
        404:
          description: Product not found
        401:
          description: Unauthorized

  /p/{id}-{slug}:
    get:
      tags: [Public - Products]
      summary: Get product by ID and slug (Self-healing URL)
      description: Public endpoint with self-healing capability - redirects if slug is incorrect
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Product ID
        - in: path
          name: slug
          required: true
          schema:
            type: string
          description: Product slug
      responses:
        200:
          description: Product found
        301:
          description: Redirect to correct URL
        404:
          description: Product not found

  /categories:
    get:
      tags: [Admin - Categories]
      summary: Get all categories
      description: Get all categories (Admin only)
      security:
        - bearerAuth: []
      responses:
        200:
          description: Categories retrieved successfully
        401:
          description: Unauthorized

    post:
      tags: [Admin - Categories]
      summary: Create a new category
      description: Create a new category (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        201:
          description: Category created successfully
        401:
          description: Unauthorized

  /categories/{id}:
    get:
      tags: [Admin - Categories]
      summary: Get category by ID
      description: Get a specific category by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Category ID
      responses:
        200:
          description: Category found
        404:
          description: Category not found
        401:
          description: Unauthorized

    put:
      tags: [Admin - Categories]
      summary: Update a category
      description: Update an existing category (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Category ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        200:
          description: Category updated successfully
        404:
          description: Category not found
        401:
          description: Unauthorized

    delete:
      tags: [Admin - Categories]
      summary: Delete a category
      description: Delete a category (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Category ID
      responses:
        200:
          description: Category deleted successfully
        404:
          description: Category not found
        401:
          description: Unauthorized

  /tags:
    get:
      tags: [Admin - Tags]
      summary: Get all tags
      description: Get all tags (Admin only)
      security:
        - bearerAuth: []
      responses:
        200:
          description: Tags retrieved successfully
        401:
          description: Unauthorized

    post:
      tags: [Admin - Tags]
      summary: Create a new tag
      description: Create a new tag (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        201:
          description: Tag created successfully
        401:
          description: Unauthorized

  /tags/{id}:
    get:
      tags: [Admin - Tags]
      summary: Get tag by ID
      description: Get a specific tag by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Tag ID
      responses:
        200:
          description: Tag found
        404:
          description: Tag not found
        401:
          description: Unauthorized

    put:
      tags: [Admin - Tags]
      summary: Update a tag
      description: Update an existing tag (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Tag ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        200:
          description: Tag updated successfully
        404:
          description: Tag not found
        401:
          description: Unauthorized

    delete:
      tags: [Admin - Tags]
      summary: Delete a tag
      description: Delete a tag (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Tag ID
      responses:
        200:
          description: Tag deleted successfully
        404:
          description: Tag not found
        401:
          description: Unauthorized